MVAR-95 Optimize DB model

+++ grails-app/domain/org/jax/mvarcore/MvarStats.groovy
@@ -9,8 +9,8 @@ class MvarStats {
     int variantCount
     int variantCanonIdentifierCount
 
-    int variantStrainCount
-    int variantTranscriptCount
+    Long variantStrainCount
+    Long variantTranscriptCount
 
     int strainAnalysisCount
     int geneAnalysisCount
